Move phone locking mobile UI to phonelocking/mobile
