+++ phonelocking/mobile/build.gradle.kts
@@ -18,7 +18,30 @@ kotlin {
         val androidMain by getting {
             dependencies {
                 implementation(libs.koin.android)
+                implementation(libs.koin.compose)
+                implementation(libs.bundles.compose.mobile)
             }
+        }
+    }
+}
+
+// Workaround for https://youtrack.jetbrains.com/issue/KT-38694
+configurations {
+    create("composeCompiler") {
+        isCanBeConsumed = false
+    }
+}
+dependencies {
+    "composeCompiler"("androidx.compose.compiler:compiler:${libs.versions.composeCompiler.get()}")
+}
+android {
+    afterEvaluate {
+        val composeCompilerJar = configurations["composeCompiler"]
+            .resolve()
+            .singleOrNull()
+            ?: error("Missing Compose compiler")
+        tasks.withType<org.jetbrains.kotlin.gradle.tasks.KotlinCompile> {
+            kotlinOptions.freeCompilerArgs += listOf("-Xuse-ir", "-Xplugin=$composeCompilerJar")
         }
     }
 }
