+++ phonelocking/mobile/src/androidMain/res/values/strings.xml
@@ -9,6 +9,8 @@
     <string name="phone_locking_accessibility_service_label">Phone Locking</string>
     <string name="phone_locking_setup_accessibility_title">Enable Accessibility Service</string>
     <string name="phone_locking_setup_accessibility_desc">Phone Locking makes use of an Accessibility Service to lock your phone. Phone Locking cannot function without this Accessibility Service. By enabling this Accessibility Service, Smartwatch Extensions will be able to lock your phone.</string>
+    <string name="phone_locking_setup_open_settings">Open Settings</string>
+    <string name="phone_locking_setup_cancel">Cancel</string>
     <string name="phone_locking_accessibility_service_desc">Allows you to lock your phone from your watch with Smartwatch Extensions. Smartwatch Extensions cannot perform any other operations through this Accessibility Service</string>
 
 </resources>
